% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_db.R
\name{add_db}
\alias{add_db}
\title{Add a series to a database}
\usage{
add_db(db_obj, series)
}
\arguments{
\item{db_obj}{db-obj.}

\item{series}{series-obj.}
}
\value{
db-obj
}
\description{
Add a series to a database
}
