% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_tex.R
\name{create_tex}
\alias{create_tex}
\title{Create a new Sweave file for database documentation}
\usage{
create_tex(db_obj, filename, title, author, version)
}
\arguments{
\item{db_obj}{A db-obj.}

\item{filename}{A character.}

\item{title}{A character.}

\item{author}{A character.}

\item{version}{A character.}
}
\value{
A character.
}
\description{
Create a new Sweave file for database documentation
}
