% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OO_classes.R
\docType{class}
\name{db-obj-class}
\alias{db-obj-class}
\title{Class definition for database object}
\value{
db-obj
}
\description{
Class definition for database object
}
\section{Slots}{

\describe{
\item{\code{series}}{list.}

\item{\code{data}}{mts.}

\item{\code{db}}{matrix.}
}}

