% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OO_classes.R
\docType{class}
\name{series-obj-class}
\alias{series-obj-class}
\title{Class definition for series object}
\value{
series-obj
}
\description{
Class definition for series object
}
\section{Slots}{

\describe{
\item{\code{mnemonic}}{character.}

\item{\code{description}}{character.}

\item{\code{units}}{character.}

\item{\code{source}}{character.}

\item{\code{table}}{character.}

\item{\code{index}}{ANY.}

\item{\code{model_desc}}{character.}

\item{\code{frequency}}{numeric.}

\item{\code{start_date}}{numeric.}

\item{\code{end_date}}{numeric.}

\item{\code{raw_data}}{ts.}

\item{\code{data}}{mts.}
}}

